Use drive name instead of id in export path
