+++ src/internal/m365/export.go
@@ -58,6 +58,7 @@ func (ctrl *Controller) ProduceExportCollections(
 			exportCfg,
 			opts,
 			dcs,
+			ctrl.backupDriveIDNames,
 			deets,
 			errs)
