+++ src/internal/m365/service/onedrive/export_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/alcionai/corso/src/internal/data"
 	dataMock "github.com/alcionai/corso/src/internal/data/mock"
-	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive"
 	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
 	odStub "github.com/alcionai/corso/src/internal/m365/service/onedrive/stub"
 	"github.com/alcionai/corso/src/internal/tester"
@@ -28,45 +28,6 @@ type ExportUnitSuite struct {
 
 func TestExportUnitSuite(t *testing.T) {
 	suite.Run(t, &ExportUnitSuite{Suite: tester.NewUnitSuite(t)})
-}
-
-func (suite *ExportUnitSuite) TestIsMetadataFile() {
-	table := []struct {
-		name          string
-		id            string
-		backupVersion int
-		isMeta        bool
-	}{
-		{
-			name:          "legacy",
-			backupVersion: version.OneDrive1DataAndMetaFiles,
-			isMeta:        false,
-		},
-		{
-			name:          "metadata file",
-			backupVersion: version.OneDrive3IsMetaMarker,
-			id:            "name" + metadata.MetaFileSuffix,
-			isMeta:        true,
-		},
-		{
-			name:          "dir metadata file",
-			backupVersion: version.OneDrive3IsMetaMarker,
-			id:            "name" + metadata.DirMetaFileSuffix,
-			isMeta:        true,
-		},
-		{
-			name:          "non metadata file",
-			backupVersion: version.OneDrive3IsMetaMarker,
-			id:            "name" + metadata.DataFileSuffix,
-			isMeta:        false,
-		},
-	}
-
-	for _, test := range table {
-		suite.Run(test.name, func() {
-			assert.Equal(suite.T(), test.isMeta, isMetadataFile(test.id, test.backupVersion), "is metadata")
-		})
-	}
 }
 
 type finD struct {
@@ -88,66 +49,6 @@ func (fd finD) FetchItemByName(ctx context.Context, name string) (data.Item, err
 	}
 
 	return nil, assert.AnError
-}
-
-func (suite *ExportUnitSuite) TestGetItemName() {
-	table := []struct {
-		tname         string
-		id            string
-		backupVersion int
-		name          string
-		fin           data.FetchItemByNamer
-		errFunc       assert.ErrorAssertionFunc
-	}{
-		{
-			tname:         "legacy",
-			id:            "name",
-			backupVersion: version.OneDrive1DataAndMetaFiles,
-			name:          "name",
-			errFunc:       assert.NoError,
-		},
-		{
-			tname:         "name in filename",
-			id:            "name.data",
-			backupVersion: version.OneDrive4DirIncludesPermissions,
-			name:          "name",
-			errFunc:       assert.NoError,
-		},
-		{
-			tname:         "name in metadata",
-			id:            "id.data",
-			backupVersion: version.Backup,
-			name:          "name",
-			fin:           finD{id: "id.meta", name: "name"},
-			errFunc:       assert.NoError,
-		},
-		{
-			tname:         "name in metadata but error",
-			id:            "id.data",
-			backupVersion: version.Backup,
-			name:          "",
-			fin:           finD{err: assert.AnError},
-			errFunc:       assert.Error,
-		},
-	}
-
-	for _, test := range table {
-		suite.Run(test.tname, func() {
-			t := suite.T()
-
-			ctx, flush := tester.NewContext(t)
-			defer flush()
-
-			name, err := getItemName(
-				ctx,
-				test.id,
-				test.backupVersion,
-				test.fin)
-			test.errFunc(t, err)
-
-			assert.Equal(t, test.name, name, "name")
-		})
-	}
 }
 
 type mockRestoreCollection struct {
@@ -391,11 +292,7 @@ func (suite *ExportUnitSuite) TestGetItems() {
 			ctx, flush := tester.NewContext(t)
 			defer flush()
 
-			ec := exportCollection{
-				baseDir:           "",
-				backingCollection: test.backingCollection,
-				backupVersion:     test.version,
-			}
+			ec := drive.NewExportCollection("", test.backingCollection, test.version)
 
 			items := ec.Items(ctx)
